+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Update class data to Factorio 1.1.12
 * Fixed "attempt to index local 'lastframe' (a nil value)" when calling `__DebugAdapter.breakpoint`
+* Use Command text as source when available
+* Disassemble Lua if no source is available
 
 ## 1.1.3